# .context/openapi.yaml
openapi: 3.0.3
info:
  title: SISE - Multimodal Retrieval & Privacy-Aware Search API
  description: >
    Production-ready contract for SISE. v1.0.0 includes:
    - flexible vector dimensions with server-side validation and explicit error codes,
    - recursive structured filters with canonical examples,
    - presigned upload flow with idempotency semantics,
    - standardized error responses and operation examples,
    - dependency-aware health probes exposing expected vector dimension.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development gateway (FastAPI)
security:
  - BearerAuth: []
paths:
...
  /media/upload-url: # Request a presigned URL for direct upload to MinIO. Endpoint này yêu cầu xác thực (Bearer token) và cho phép client yêu cầu một URL có chữ ký để tải tệp trực tiếp lên MinIO mà không cần đi qua backend. Client cung cấp filename, content_type và optional expected_size_mb để server có thể tạo URL phù hợp với chính sách lưu trữ và bảo mật.
    post:
      operationId: requestUploadUrl
      summary: Request a presigned URL for direct upload to MinIO
      tags: [Media]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, content_type]
              properties:
                filename: { type: string, example: "vacation.jpg" } # Đây là tên tệp mà client dự định tải lên. Server có thể sử dụng thông tin này để tạo object key trong MinIO hoặc áp dụng chính sách đặt tên.
                content_type: { type: string, example: "image/jpeg" } # Đây là loại nội dung của tệp. Server có thể sử dụng thông tin này để xác định cách xử lý tệp hoặc áp dụng chính sách bảo mật.
                expected_size_mb:
                  type: integer
                  description: Optional hint of expected file size in MB (server may validate)
            examples:
              request:
                value:
                  filename: "vacation.jpg"
                  content_type: "image/jpeg"
                  expected_size_mb: 5
      responses:
        '200': # Status 200: OK, có nghĩa là URL có chữ ký đã được tạo thành công và trả về trong phản hồi. Client có thể sử dụng URL này để tải tệp trực tiếp lên MinIO. Phản hồi bao gồm upload_url (URL có chữ ký), object_key (khóa đối tượng trong MinIO), expires_in_sec (thời gian hiệu lực của URL), max_file_size_mb (kích thước tệp tối đa được phép) và allowed_content_types (danh sách loại nội dung được phép).
          description: Presigned URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedUploadResponse'
        '400': # Status 400: Bad Request, có nghĩa là dữ liệu đầu vào không hợp lệ, ví dụ thiếu trường bắt buộc hoặc định dạng sai.
          $ref: '#/components/responses/StandardError'
        '401': # Status 401: Unauthorized, có nghĩa là yêu cầu không có token hợp lệ hoặc token đã hết hạn.
          $ref: '#/components/responses/Unauthorized'
        '409': # Status 409: Conflict, có nghĩa là có xung đột với yêu cầu hiện tại, ví dụ idempotency key đã tồn tại và đang được xử lý. Đây là tín hiệu cho client rằng yêu cầu này có thể đã được xử lý trước đó và nên tránh gửi lại.
          description: Duplicate idempotent request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
  # ------------------------------------------------------------------------------------------------------------------
  /media/upload: # Legacy multipart upload proxied through backend (fallback). Endpoint này cũng cho phép tải tệp lên thông qua backend bằng cách gửi multipart/form-data. Đây là phương thức cũ và không được khuyến khích, nhưng vẫn tồn tại như một fallback nếu client không thể sử dụng presigned URL. Client gửi tệp cùng với metadata như album_id và privacy_level. Server xử lý tải lên, lưu trữ tệp vào MinIO, và trả về metadata của ảnh đã lưu.
    post:
      operationId: uploadLegacy
      summary: Legacy multipart upload proxied through backend (fallback)
      tags: [Media]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                album_id: { type: integer }
                privacy_level: { $ref: '#/components/schemas/PrivacyLevel' }
            encoding:
              file:
                contentType: image/*
      responses:
        '201': # Status 201: Created, có nghĩa là tệp đã được tải lên thành công và metadata đã được đăng ký. Phản hồi bao gồm image_id (UUID của ảnh), minio_url (URL truy cập ảnh trong MinIO), status (trạng thái hiện tại của ảnh, ví dụ "pending" nếu đang chờ xử lý), và index_status (trạng thái của việc lập chỉ mục vào vector DB, ví dụ "pending" hoặc "ready"). Lưu ý rằng lập chỉ mục có thể là một quá trình bất đồng bộ, vì vậy ngay cả khi tệp đã được tải lên thành công, nó có thể vẫn đang chờ để được lập chỉ mục.
          description: File uploaded and metadata registered (indexing may be pending)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400': # Status 400: Bad Request, có nghĩa là dữ liệu đầu vào không hợp lệ, ví dụ thiếu trường bắt buộc, định dạng sai, hoặc tệp không hợp lệ.
          $ref: '#/components/responses/StandardError'
        '401': # Status 401: Unauthorized, có nghĩa là yêu cầu không có token hợp lệ hoặc token đã hết hạn.
          $ref: '#/components/responses/Unauthorized'
        '413': # Status 413: Payload Too Large, có nghĩa là tệp được gửi vượt quá giới hạn kích thước tối đa mà server cho phép. Phản hồi sẽ bao gồm mã lỗi và thông điệp giải thích rằng tệp quá lớn.
          description: Payload too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /media/upload/confirm: # Confirm and finalize an upload (completes S5 step). Endpoint này yêu cầu xác thực và cho phép client xác nhận rằng file đã được tải lên thành công và sẵn sàng để được xử lý. Đây là bước cuối (S5) trong quy trình tải lên. Client gửi object_key từ presigned upload, server sẽ verify file tồn tại trên MinIO, lưu metadata vào PostgreSQL, và queue file cho việc embedding và indexing.
    post:
      operationId: confirmUpload
      summary: Confirm and finalize an upload (completes S5 step)
      tags: [Media]
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [object_key, album_id, privacy_level]
              properties:
                object_key: { type: string, description: "The MinIO object key from presigned upload" }
                album_id: { type: integer }
                privacy_level: { $ref: '#/components/schemas/PrivacyLevel' }
                tags:
                  type: array
                  items: { type: string }
                  description: Optional tags for the uploaded image
      responses:
        '200': # Status 200: OK, có nghĩa là file đã được xác nhận thành công. Server sẽ bắt đầu process embedding và indexing.
          description: Upload confirmed and queued for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400': # Status 400: Bad Request, có nghĩa là dữ liệu đầu vào không hợp lệ.
          $ref: '#/components/responses/StandardError'
        '401': # Status 401: Unauthorized, có nghĩa là yêu cầu không có token hợp lệ hoặc token đã hết hạn.
          $ref: '#/components/responses/Unauthorized'

    get:
      operationId: getMediaMetadata
      summary: Retrieve image metadata and signed GET URL (if authorized)
      tags: [Media]
      security:
        - BearerAuth: []
      parameters:
        - name: image_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200': # Status 200: OK, có nghĩa là metadata của ảnh đã được truy xuất thành công và trả về trong phản hồi. Phản hồi bao gồm image_id (UUID của ảnh), user_id (ID người sở hữu ảnh), album_id (ID album chứa ảnh), minio_url (URL có chữ ký để truy cập ảnh trong MinIO), privacy_level (mức độ riêng tư của ảnh), tags (danh sách thẻ liên quan đến ảnh), created_at (thời gian tạo ảnh), và index_status (trạng thái lập chỉ mục vào vector DB). Lưu ý rằng URL có chữ ký chỉ được trả về nếu client có quyền truy cập dựa trên metadata của ảnh.
          description: Image metadata and access URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageMetadata'
        '403': # Status 403: Forbidden, có nghĩa là client không có quyền truy cập vào ảnh này dựa trên metadata (ví dụ privacy_level hoặc user_id không khớp). Phản hồi sẽ bao gồm mã lỗi và thông điệp giải thích rằng quyền truy cập bị từ chối.
          $ref: '#/components/responses/StandardError'
        '404': # Status 404: Not Found, có nghĩa là không tìm thấy ảnh với image_id được cung cấp. Phản hồi sẽ bao gồm mã lỗi và thông điệp giải thích rằng ảnh không tồn tại.
          $ref: '#/components/responses/StandardError'
        '401': # Status 401: Unauthorized, có nghĩa là yêu cầu không có token hợp lệ hoặc token đã hết hạn.
          $ref: '#/components/responses/Unauthorized'
  /media/{image_id}/update: # Update image metadata (album, privacy, tags). Endpoint này yêu cầu xác thực và cho phép owner của ảnh cập nhật metadata như album_id, privacy_level, hoặc tags.
    put:
      operationId: updateMediaMetadata
      summary: Update image metadata (album, privacy, tags)
      tags: [Media]
      security:
        - BearerAuth: []
      parameters:
        - name: image_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                album_id: { type: integer }
                privacy_level: { $ref: '#/components/schemas/PrivacyLevel' }
                tags:
                  type: array
                  items: { type: string }
      responses:
        '200': # Status 200: OK, có nghĩa là metadata đã được cập nhật thành công.
          description: Image metadata updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageMetadata'
        '400': # Status 400: Bad Request, có nghĩa là dữ liệu đầu vào không hợp lệ.
          $ref: '#/components/responses/StandardError'
        '401': # Status 401: Unauthorized, có nghĩa là yêu cầu không có token hợp lệ hoặc token đã hết hạn.
          $ref: '#/components/responses/Unauthorized'
        '403': # Status 403: Forbidden, có nghĩa là client không phải là owner của ảnh này.
          $ref: '#/components/responses/StandardError'
        '404': # Status 404: Not Found, có nghĩa là không tìm thấy ảnh.
          $ref: '#/components/responses/StandardError'
  /media/{image_id}/delete: # Delete (soft delete) an image. Endpoint này yêu cầu xác thực và cho phép owner của ảnh xóa (soft delete) ảnh.
    delete:
      operationId: deleteMedia
      summary: Delete (soft delete) an image
      tags: [Media]
      security:
        - BearerAuth: []
      parameters:
        - name: image_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204': # Status 204: No Content, có nghĩa là ảnh đã được xóa thành công.
          description: Image deleted successfully
        '401': # Status 401: Unauthorized, có nghĩa là yêu cầu không có token hợp lệ hoặc token đã hết hạn.
          $ref: '#/components/responses/Unauthorized'
        '403': # Status 403: Forbidden, có nghĩa là client không phải là owner của ảnh này.
          $ref: '#/components/responses/StandardError'
        '404': # Status 404: Not Found, có nghĩa là không tìm thấy ảnh.
          $ref: '#/components/responses/StandardError'
  /media: # List images for the authenticated user. Endpoint này yêu cầu xác thực và cho phép user lấy danh sách các ảnh của chính mình.
    get:
      operationId: listMedia
      summary: List images for the authenticated user
      tags: [Media]
      security:
        - BearerAuth: []
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: album_id
          in: query
          schema:
            type: integer
          description: Filter by album ID
      responses:
        '200': # Status 200: OK, có nghĩa là danh sách ảnh đã được truy xuất thành công.
          description: List of images
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ImageMetadata'
                  total: { type: integer }
                  offset: { type: integer }
                  limit: { type: integer }
        '401': # Status 401: Unauthorized, có nghĩa là yêu cầu không có token hợp lệ hoặc token đã hết hạn.
          $ref: '#/components/responses/Unauthorized'

  # --- AI & INFERENCE ----------------------------------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: >
        Unique key to prevent duplicate processing of the same upload/index request.
        Server stores idempotency keys for a configurable TTL (recommended 24h). Keys are scoped per authenticated user.
  responses:
    StandardError:
      description: Standard error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    PrivacyLevel: # Privacy level of media (0: Private, 1: Friends, 2: Public)
      type: integer
      enum: [0, 1, 2]
      description: "0: Private, 1: Friends, 2: Public"
...
    PresignedUploadResponse: # Response schema for presigned upload URL endpoint
      type: object
      properties:
        upload_url:
          type: string
          format: uri
        object_key:
          type: string
        expires_in_sec:
          type: integer
          example: 3600
        max_file_size_mb:
          type: integer
          example: 20
        allowed_content_types:
          type: array
          items:
            type: string
          example: ["image/jpeg", "image/png"]
        note:
          type: string
          description: "Client should upload within expires_in_sec and respect allowed_content_types."
    UploadResponse: # Response schema for media upload endpoint (legacy)
      type: object
      properties:
        image_id:
          type: string
          format: uuid
        minio_url:
          type: string
          format: uri
        status:
          type: string
          example: "pending"
        index_status:
          type: string
          example: "pending"
    ImageMetadata: # Metadata associated with an image, used for access control and search filtering
      type: object
      properties:
        image_id: { type: string, format: uuid }
        user_id: { type: integer }
        album_id: { type: integer }
        minio_url: { type: string, format: uri }
        privacy_level: { $ref: '#/components/schemas/PrivacyLevel' }
        tags:
          type: array
          items: { type: string }
        created_at:
          type: string
          format: date-time
        index_status:
          type: string
          example: "ready"

    # --- SEARCH RESPONSE --------------------------------------
...
    Error: # Standard error response schema for all endpoints
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "ERR_VECTOR_DIM_MISMATCH"
        message:
          type: string
          example: "Expected 512 items, received 768"
        details:
          type: object
          description: Optional structured details for debugging
...